Added Python solution to Easy#30
I wanted something little bit different than a straight forward double for and try all numbers.
I start by removing all numbers greater than the target and sort the resulting list. I sorted it to be able to use the pop() method which is O(1) compared to remove which is O(n).
And then I use a deque to add **all** two numbers that sum up to the target.